+++ mime/test.ts
@@ -1,2 +1,2 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-import "./formfile_test.ts";
+import "./multipart_test.ts";
